# [hashdisk] On same key, return instead of override key