Since we are by definition immutable, we could just return instead of override on already seen keys.
This should give us a perf boost but prevent us from overriding a bad key