@Viq111 2018-12-18
<details>
 <summary>Benchmarks results </summary>

```gf
name               old time/op     new time/op     delta
HashDiskWrite-2       5.21µs ± 2%     5.34µs ± 2%    ~     (p=0.063 n=5+4)
HashDiskRead-2        1.23µs ± 1%     1.56µs ±30%    ~     (p=0.063 n=5+5)
KvimdRandbo-2          346ns ± 1%      357ns ± 4%  +3.24%  (p=0.008 n=5+5)
KvimdWrite-2          6.15µs ± 1%     6.51µs ± 1%  +5.84%  (p=0.008 n=5+5)
KvimdReadSame-2        152ns ±20%      153ns ±20%    ~     (p=0.810 n=5+5)
KvimdReadRandom-2    10.0µs ±210%    13.3µs ±289%    ~     (p=1.000 n=5+5)
ValuesDiskSet-2       34.5ns ± 0%     34.6ns ± 1%    ~     (p=0.905 n=5+5)
ValuesDiskGet-2       32.9ns ± 1%     33.1ns ± 0%    ~     (p=0.135 n=5+5)

name               old speed       new speed       delta
HashDiskWrite-2     4.61MB/s ± 2%   4.39MB/s ±10%  -4.77%  (p=0.024 n=5+5)
HashDiskRead-2      19.5MB/s ± 1%   16.2MB/s ±27%    ~     (p=0.063 n=5+5)
KvimdRandbo-2        335MB/s ± 1%    324MB/s ± 4%  -3.13%  (p=0.008 n=5+5)
KvimdWrite-2        18.9MB/s ± 1%   17.8MB/s ± 1%  -5.51%  (p=0.008 n=5+5)
KvimdReadSame-2      705MB/s ±64%    701MB/s ±64%    ~     (p=1.000 n=5+5)
KvimdReadRandom-2  40.3MB/s ±164%  41.7MB/s ±120%    ~     (p=1.000 n=5+5)
ValuesDiskSet-2      232MB/s ± 0%    231MB/s ± 1%    ~     (p=1.000 n=5+5)
ValuesDiskGet-2      243MB/s ± 1%    241MB/s ± 1%    ~     (p=0.222 n=5+5)
```