@Viq111 2018-12-18
<details>
 <summary>Extended benchmarks results </summary>

```gf
name                           old time/op    new time/op    delta
HashDiskSetWithLoad_0.00-0.50     496ns ± 0%    1115ns ± 0%   ~     (p=1.000 n=1+1)
HashDiskSetWithLoad_0.70-0.90    1.04µs ± 0%    1.04µs ± 0%   ~     (p=1.000 n=1+1)
HashDiskSetWithLoad_0.90-0.95    2.97µs ± 0%    2.63µs ± 0%   ~     (p=1.000 n=1+1)
HashDiskSetWithLoad_0.95-0.99    17.3µs ± 0%    16.9µs ± 0%   ~     (p=1.000 n=1+1)
HashDiskGetWithLoad_0.00-0.50     292ns ± 0%     278ns ± 0%   ~     (p=1.000 n=1+1)
HashDiskGetWithLoad_0.70-0.90     665ns ± 0%     656ns ± 0%   ~     (p=1.000 n=1+1)
HashDiskGetWithLoad_0.90-0.95    2.03µs ± 0%    2.01µs ± 0%   ~     (p=1.000 n=1+1)
HashDiskGetWithLoad_0.95-0.99    15.9µs ± 0%    15.9µs ± 0%   ~     (p=1.000 n=1+1)

name                           old speed      new speed      delta
HashDiskSetWithLoad_0.00-0.50  48.3MB/s ± 0%  21.5MB/s ± 0%   ~     (p=1.000 n=1+1)
HashDiskSetWithLoad_0.70-0.90  23.1MB/s ± 0%  23.2MB/s ± 0%   ~     (p=1.000 n=1+1)
HashDiskSetWithLoad_0.90-0.95  8.08MB/s ± 0%  9.11MB/s ± 0%   ~     (p=1.000 n=1+1)
HashDiskSetWithLoad_0.95-0.99  1.39MB/s ± 0%  1.42MB/s ± 0%   ~     (p=1.000 n=1+1)
HashDiskGetWithLoad_0.00-0.50  81.9MB/s ± 0%  86.3MB/s ± 0%   ~     (p=1.000 n=1+1)
HashDiskGetWithLoad_0.70-0.90  36.1MB/s ± 0%  36.6MB/s ± 0%   ~     (p=1.000 n=1+1)
HashDiskGetWithLoad_0.90-0.95  11.8MB/s ± 0%  11.9MB/s ± 0%   ~     (p=1.000 n=1+1)
HashDiskGetWithLoad_0.95-0.99  1.51MB/s ± 0%  1.51MB/s ± 0%   ~     (all equal)
```