@Viq111 2018-12-18
Doesn't seem to improve performance, so let's keep it as overriding instead